# Improve TileLayer::parseCSV performance