I was trying tmxlite on a biggish map (mostly just to see how my loading code handles it), and it was super slow. After a bit of profiling I figured out that the stringstream in parseCSV is taking up most of the runtime (unfortunately the way iostreams is designed, it very hard to implement it with good performace), so I replaced it with `std::strtoull` (should be part of C++11, but I haven't tested this code on anything but Linux). As an extra, I also added a reserve to the vector, it also helps a bit. (By the way, shouldn't we expose the width and height of layers too?) I'm copy-pasting the measurements result from my commit message:

Performance measurements (time to run ParseTest) on a 4096x4096 map with two layers:

* clang with libc++:
  original: 6.209s, strtoull: 0.773s, strtoull+reserve: 0.723s
* clang with libstdc++:
  original: 2.153s, strtoull: 0.800s, strtoull+reserve: 0.736s
* gcc:
  original: 2.115s, strtoull: 0.805s, strtoull+reserve: 0.752s

Unfortunately, libc++'s iostream is especially crap compared to libstdc++'s, here this simple change achieved about 8x speedup, with gcc it's only 2.5x faster. The above numbers are with statically linked tmxlite compiled in release mode. (It's worse in my app where I compiled the code with asan and ubsan (including libc++), the optimized version takes about 14s, the original is unbearably slow.)