@fallahn 2020-05-06
Thanks for taking the time to check this out! I haven't used CSV formatted files personally, so I hadn't come up against this before. I'd like to compare this on windows too, but I don't see why the case would be any different there. And you're right, the size of layers should be exposed too. I think I originally skipped this as they were always the same size as the map, but infinite maps have since been introduced making layer sizes now necessary.