@u3shit 2020-05-08
Thanks, I originally used the CSV format because my old importer script could only handle that format, but now with tmxlite I could probably change that. The one advantage of csv format is that it's at least a bit human readable-ish in git diffs.
Anyway, I've created a new pull request to expose the size of the layers.